init.target.rc: wait for adsp_loader before booting ADSP
With GKI, almost all the drivers are loaded as modules and when
init script writes to "/sys/kernel/boot_adsp/boot", adsp_loader_dlkm
is not even loaded thereby not bringing ADSP out of reset. However,
this is required for battery charger driver to probe to support
charger mode. Hence wait for "/sys/kernel/boot_adsp/boot" which is
a must before writing to it. Since QTI health HAL implementation
already waits for the presence of battery power supply device, remove
waiting for it in init.target.rc.

Change-Id: Ib471573a80563a48bd859c118be6fb3683bb58a0